+++ src/plugins/worker/workers/webpack.config.js
@@ -49,7 +49,7 @@ module.exports = {
                 use: {
                     loader: 'babel-loader',
                     options: babelConfig
-                }
+                },
             }
         ],
     },