+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 <!-- CHANGELOG NEWEST -->
 
+- IMPROVE usage of the [mingo query library](https://github.com/kofrasa/mingo) by only adding the operators to the build that are really needed.
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->